Revert "Hit Enter to validate" for mainFolder and configFile
Bugs appear with this logic.
To reproduce:
- enter a valid configFile and hit enter, that will populate
the param comboboxes.
- modify the configFile so that it's an invalid path, but don't
press Enter so that the focus stays in the configFile lineEdit.
- if you now click on a parameter combobox or try to switch it with
the mouse wheel, it will first remove focus from the lineEdit, which
triggers the signal editingFinished, which calls configFile_changed(),
which deletes the comboboxes since the file is not valid.
- then it crashes because you're trying to access a non-existing combobox.

We keep the "Hit Enter to validate" for filePattern, because it's the one
which brought motivation to change from textChanged to editingFinished, and
because changing it doesn't delete any widgets.
I added a place holder text to tell the user about this behaviour