Add support for glob pattern and index selector in file pattern
The file pattern must be something like: 'image_*.png[<idx>]'
You can replace <idx> with a number that selects which one of the matches
to select. Negative indices count from the end: -1 is the last one.
The files that are matched are ordered in alphanumerical order.
The matched pattern is drawn in the corner of each image.