Allow experiments to be run in any order when using only_exp_id
This is useful when running experiments as a job array on a cluster.
Most of the work was handling the CSV correctly, and allowing to write
in a header-less CSV, and pre-pending one later on, when we know the
number of results that the experiment returns.
- The problem was that if we started with redundant experiments,
we don't know in advance the number of results that the experiment
returns. So we just don't write the result for the redundant
experiments anymore. This is now handled by the viewer, since it
has the src_exp_id.
- Make a function that takes care of creating/prepending the header
- Fix error of symlink already existing
- Factor code for the prefix values in the csv row (exp_id, src_exp_id
 and parameters)
- Remove the result_list since it's not used anymore.

In the viewer:
- Do not populate result combobox if no CSV was read
- Impute redundant values in CSV to create the resultArray.